Query: Fix for GroupBy with Where Aggregate produces incorrect SQL
Issue: Due to the shape of QM we incorrectly identified the query for Group By SQL translation.
Which is not the case hence we generate a SQL which produces incorrect result
Fix: During result operator processing, we more carefully identify if we are selecting from a grouping source.
Also relax criterea that group by aggregate should be only 1 level deep. This also allows us to translate group by aggregate when it is inside join

Resolves #11695